Componentize html and start to add redux
