+++ package.json
@@ -26,6 +26,8 @@
     "jquery": "^3.2.1",
     "react": "^15.5.4",
     "react-dom": "^15.5.4",
+    "react-redux": "^5.0.4",
+    "redux": "^3.6.0",
     "webpack-dev-server": "^2.4.2"
   }
 }
